Table views at Table and Update edited

+++ Models/Product.cs
@@ -6,10 +6,10 @@ namespace Food_Registration.Models
     public int ProductId { get; set; }
     public string Name { get; set; } = string.Empty;
     public string? Category { get; set; }
-    public int? Calorie {get; set; }
-    public int? Carbonhydrate {get; set; }
-    public int? Fat {get; set; }
-    public int? Protein {get; set; } 
+    public decimal? Calorie {get; set; }
+    public decimal? Carbonhydrate {get; set; }
+    public decimal? Fat {get; set; }
+    public decimal? Protein {get; set; } 
     public string? Description { get; set; }
     public string? ImageUrl { get; set; }
   }
